added Login and SignUp funcionality. Fixed updateAverageStars in Filmography.java. Fixed miniproblems to make it work. Added some functions to AccountRepo and service. Added automatic ID to Account

+++ practica2/src/main/java/es/code/urjc/practica2/model/Filmography.java
@@ -86,7 +86,8 @@ public class Filmography {
         if (filmographyReviews.isEmpty()) {
             this.filmographyAverageStars = 0;
         } else {
-            this.filmographyAverageStars = (float) filmographyReviews.stream().mapToInt(Review::getReviewStars).average().orElse(0.0);
+
+            this.filmographyAverageStars = (float) filmographyReviews.stream().mapToDouble(Review::getReviewStars).average().orElse(0.0);
         }
     }
